requirements.txt
requests==2.32.3
ippanel==2.0.7
python-dotenv==1.0.1